requirements.txt
unstructured[pdf]==0.16.5
langchain==0.3.7
langchain-community==0.3.7
langchain-groq== 0.2.1
langchain_core==0.3.19
langchain_chroma==0.1.4
groq==0.12.0
sentence-transformers== 3.3.1
flask==3.1.0
Flask_Cors==5.0.0

python-dotenv

pytesseract==0.3.13
mysql-connector-python==9.1.0